Changed percision to double and modified tests. Sigmoid activation now returns a double value instead of a vector.

+++ CPUNeuralNetwork/CPUTesting.hpp
@@ -17,7 +17,7 @@ namespace cpu {
 
         // Helper functions
 
-        static bool areFloatEqual(float a, float b);
+        static bool areFloatEqual(double a, double b);
 
     };
 }
